Added additional functionality
fixed minor issues with logic and stuff,
also add a basic(non finished) output for the field
The key isn't printing correctly so that NEEDS to be fixed. Also the
words aren't in the playing field so that's an issue. I believe the
issue is in the aweful and huge for loop.